+++ CHANGELOG.md
@@ -13,6 +13,13 @@
 * Add the possibility to correctly delete fixed due date schedule range. Refs UICIRC-1026.
 * Cover Widget by jest tests. Refs UICIRC-899.
 * Add page titles to Circulation settings pages. Refs UICIRC-970.
+* RulesEditor should use stripes.hasPerm. Refs UICIRC-1047.
+* Disable block template under reminder fees. Refs UICIRC-1046.
+* UI tests replacement with RTL/Jest for `CirculationRules.js`. Refs UICIRC-958.
+
+## [9.0.4](https://github.com/folio-org/ui-circulation/tree/v9.0.4) (2024-02-22)
+[Full Changelog](https://github.com/folio-org/ui-circulation/compare/v9.0.3...v9.0.4)
+
 * Fix issue with 'Enter' key in circulation rules editor that adds indent to previous line. Refs UICIRC-1042.
 
 ## [9.0.3](https://github.com/folio-org/ui-circulation/tree/v9.0.3) (2023-11-09)